[bo 230919] 이벤트 기능 구현
