+++ src/main/resources/mappers/winnerMapper.xml
@@ -1,2 +1,42 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="winner">
+    <!-- 이벤트 전체 참여자 -->
+    <select id="applyList" resultType="kr.co.teaspoon.dto.Apply">
+        select * from apply where eno=#{eno}
+    </select>
+    <!-- 5명 당첨자 추첨 -->
+    <insert id="random">
+        insert into winnerList select * from apply where eno=#{eno} AND eno NOT IN (SELECT distinct eno FROM winnerList) order by rand() limit 5;
+    </insert>
+    <!-- 5명 당첨자 목록 -->
+    <select id="winners" resultType="kr.co.teaspoon.dto.WinnerList">
+        SELECT * FROM winnerList where eno=#{eno};
+    </select>
+
+    <!-- 당첨자 발표 게시글 목록 -->
+    <select id="winnerList" resultType="kr.co.teaspoon.dto.Winner">
+        select * from winner order by resdate desc
+    </select>
+    <!-- 당첨자 발표 게시글 카운트 -->
+    <select id="winnerListCount" resultType="integer">
+        select count(*) from winner
+    </select>
+    <!--당첨자 발표 게시글 상세보기 -->
+    <select id="winnerDetail" resultType="kr.co.teaspoon.dto.Winner">
+        select * from winner where wno=#{wno}
+    </select>
+    <!-- 당첨자 발표 글쓰기 -->
+    <insert id="winnerInsert">
+        insert into winner values (default, #{eno}, #{title}, #{content}, #{author}, default)
+    </insert>
+
+    <update id="winnerUpdate">
+        update winner set title=#{title}, content=#{content} where wno=#{wno}
+    </update>
+
+    <delete id="winnerDelete">
+        delete from winner where wno=#{wno}
+    </delete>
 
 </mapper>
